+++ CHANGELOG.md
@@ -1,13 +1,32 @@
 ## v0.17.0
+
+Here are some highlights of changes in this version. See the full list of changes for more details: https://github.com/jupyter-widgets/ipyleaflet/compare/0.16.0...0.17.0
+
 ### New Features
 
 * Make it possible to use Choropleth layer with data containing NaNs [#972](https://github.com/jupyter-widgets/ipyleaflet/pull/972)
 * Add Map panes [#999](https://github.com/jupyter-widgets/ipyleaflet/pull/999)
 * Allow setting Map.dragging [#1001](https://github.com/jupyter-widgets/ipyleaflet/pull/1001)
 * Add visible attribute to GeoJSON layer [#1002](https://github.com/jupyter-widgets/ipyleaflet/pull/1002)
-* [BREAKING CHANGE] Remove get and set decorators in LegendControl [#979](https://github.com/jupyter-widgets/ipyleaflet/pull/979)
 
-## Maintenance
+### Deprecated API
+
+* Deprecate LegendControl properties `name`, `legends`, `positioning`, and `positionning` [#979](https://github.com/jupyter-widgets/ipyleaflet/pull/979) and [#1005](https://github.com/jupyter-widgets/ipyleaflet/pull/1005). Update your code with the following substitutions for a LegendControl `legend`:
+  * `legend.name` -> `legend.title`
+  * `legend.legends` -> `legend.legend`
+  * `legend.positioning` -> `legend.position`
+  * `legend.positionnning` -> `legend.position`
+  
+  The `name` argument in creating a LegendControl is also deprecated, please use the `title` argument instead: `LegendControl({}, title='My Title')`.
+* Deprecate layer and control-specific method names for Map and LayerGroup, in favor of methods that work for both layers and controls [#982](https://github.com/jupyter-widgets/ipyleaflet/pull/982). Update your code with the following substitutions for a Map `map` (or LayerGroup):
+  * `map.add_control(...)` or `map.add_layer(...)` -> `map.add(...)`
+  * `map.remove_control(...)` or `map.remove_layer(...)` -> `map.remove(...)`
+  * `map.substitute_control(...)` or `map.substitute_layer(...)` -> `map.substitute(...)`
+  * `map.clear_controls(...)` or `map.clear_layers(...)` -> `map.clear(...)`
+
+  The inline operators still continue to work as before, such as `map += control` or `map -= layer`.
+
+### Maintenance
 
 * Compute the public path automatically [#988](https://github.com/jupyter-widgets/ipyleaflet/pull/988)
 
@@ -15,8 +34,6 @@
 
 * Document use of multiple basemaps [#971](https://github.com/jupyter-widgets/ipyleaflet/pull/971)
 * Add a small introduction text [#992](https://github.com/jupyter-widgets/ipyleaflet/pull/992)
-
-**Full Changelog**: https://github.com/jupyter-widgets/ipyleaflet/compare/0.16.0...0.17.0
 
 ## v0.16.0
 ### New features